Fix genMonoRepoJson
