+++ build-tools/packages/build-tools/src/index.ts
@@ -35,3 +35,4 @@ export { Timer } from "./common/timer";
 export { execAsync, execWithErrorAsync, readJsonAsync, readFileAsync, writeFileAsync } from "./common/utils";
 export { Handler } from "./repoPolicyCheck/common";
 export { policyHandlers } from "./repoPolicyCheck/handlers";
+export { generateMonoRepoInstallPackageJson } from "./genMonoRepoPackageJson/lib";
